LDE-714: removed reading tests from kartothek.io.testing package
