--- kartothek/io/testing/read.py
@@ -1,28 +1,0 @@
-"""
-This module is a collection of tests which should be implemented by all kartothek
-**read** backends. The tests are not subject to the semantic versioning scheme and may change with minor or even patch releases.
-
-To use the tests of this module, add the following import statement to your test module and ensure that the following fixtures are available in your test environment.
-
-```
-from kartothek.io.testing.read import *  # noqa
-```
-
-Fixtures required to be implemented:
-
-* ``output_type`` - One of {`dataframe`, `metpartition`, `table`} to define the outptu type of the returned result.
-* ``bound_load_dataframes`` - A callable which will retrieve the partitions in the format specified by ``output_type``. The callable should accept all keyword arguments expected for a kartothek reader.
-
-Source test data
-
-* ``dataset`` - A fixture generating test data (TODO: Expose this as a testing function)
-* ``store_factory`` - A function scoped store factory
-* ``store_session_factory`` - A session scoped store factory
-
-Feature toggles (optional):
-
-The following fixtures should be present (see tests.read.conftest)
-* ``use_categoricals`` - Whether or not the call retrievs categorical data.
-* ``dates_as_object`` - Whether or not the call retrievs date columns as objects.
-
-"""
