+++ tests/io/common/conftest.py
@@ -20,21 +20,6 @@ def output_type(request, backend_identifier):
     return request.param
 
 
-@pytest.fixture(params=[True, False], ids=["use_factory", "no_factory"])
-def use_dataset_factory(request, dates_as_object):
-    return request.param
-
-
-@pytest.fixture(params=[True, False], ids=["dates_as_object", "datest_as_datetime"])
-def dates_as_object(request):
-    return request.param
-
-
-@pytest.fixture(params=[True, False], ids=["use_categoricals", "no_categoricals"])
-def use_categoricals(request):
-    return request.param
-
-
 @pytest.fixture(scope="session")
 def dataset_dispatch_by_uuid():
     import uuid
